+++ homepage/homepage/components/forMdx.tsx
@@ -9,6 +9,7 @@ import {
 } from "@/components/docs/ContentByFramework";
 import { CodeGroup as CodeGroupClient } from "gcmp-design-system/src/app/components/molecules/CodeGroup";
 import { ComingSoon as ComingSoonClient } from "./docs/ComingSoon";
+import { IssueTrackerPreview as IssueTrackerPreviewClient } from "./docs/IssueTrackerPreview";
 
 export function CodeExampleTabs(props: CodeExampleTabsProps) {
   return <CodeExampleTabsClient {...props} />;
@@ -24,4 +25,8 @@ export function ComingSoon() {
 
 export function ContentByFramework(props: ContentByFrameworkProps) {
   return <ContentByFrameworkClient {...props}></ContentByFrameworkClient>;
+}
+
+export function IssueTrackerPreview() {
+  return <IssueTrackerPreviewClient />;
 }