+++ packages/jazz-tools/src/tests/coMap.test.ts
@@ -38,8 +38,6 @@ describe("Simple CoMap operations", async () => {
     crypto: Crypto,
   });
 
-  console.log("TestMap schema", TestMap.prototype._schema);
-
   const birthday = new Date();
 
   const map = TestMap.create(
@@ -80,6 +78,13 @@ describe("Simple CoMap operations", async () => {
     );
 
     expect(mapWithExtra.color).toEqual("red");
+  });
+
+  test("Empty schema", () => {
+    const emptyMap = CoMap.create({}, { owner: me });
+
+    // @ts-expect-error
+    expect(emptyMap.color).toEqual(undefined);
   });
 
   describe("Mutation", () => {
@@ -468,10 +473,8 @@ describe("CoMap resolution", async () => {
     const queue = new cojsonInternals.Channel<TestMap>();
 
     TestMap.subscribe(map.id, meOnSecondPeer, {}, (subscribedMap) => {
-      console.log(
-        "subscribedMap.nested?.twiceNested?.taste",
-        subscribedMap.nested?.twiceNested?.taste,
-      );
+      // Read to property to trigger loading
+      subscribedMap.nested?.twiceNested?.taste;
       void queue.push(subscribedMap);
     });