+++ packages/jazz-tools/src/tests/groupsAndAccounts.test.ts
@@ -166,68 +166,6 @@ describe("Group inheritance", () => {
     expect(mapAsReaderAfterUpdate?.title).toBe("In Grand Child");
   });
 
-  test("Group inheritance should fail if the current account doesn't have admin role in both groups", async () => {
-    const me = await Account.create({
-      creationProps: { name: "Hermes Puggington" },
-      crypto: Crypto,
-    });
-
-    const other = await Account.createAs(me, {
-      creationProps: { name: "Another user" },
-    });
-
-    const parentGroup = Group.create({ owner: me });
-    parentGroup.addMember(other, "writer");
-    const group = Group.create({ owner: me });
-    group.addMember(other, "admin");
-
-    const parentGroupOnTheOtherSide = await Group.load(
-      parentGroup.id,
-      other,
-      {},
-    );
-    const groupOnTheOtherSide = await Group.load(group.id, other, {});
-
-    if (!groupOnTheOtherSide || !parentGroupOnTheOtherSide) {
-      throw new Error("CoValue not available");
-    }
-
-    expect(() => groupOnTheOtherSide.extend(parentGroupOnTheOtherSide)).toThrow(
-      "To extend a group, the current account must have admin role in both groups",
-    );
-  });
-
-  test("Group inheritance should work if the current account has admin role in both groups", async () => {
-    const me = await Account.create({
-      creationProps: { name: "Hermes Puggington" },
-      crypto: Crypto,
-    });
-
-    const other = await Account.createAs(me, {
-      creationProps: { name: "Another user" },
-    });
-
-    const parentGroup = Group.create({ owner: me });
-    parentGroup.addMember(other, "admin");
-    const group = Group.create({ owner: me });
-    group.addMember(other, "admin");
-
-    const parentGroupOnTheOtherSide = await Group.load(
-      parentGroup.id,
-      other,
-      {},
-    );
-    const groupOnTheOtherSide = await Group.load(group.id, other, {});
-
-    if (!groupOnTheOtherSide || !parentGroupOnTheOtherSide) {
-      throw new Error("CoValue not available");
-    }
-
-    expect(() =>
-      groupOnTheOtherSide.extend(parentGroupOnTheOtherSide),
-    ).not.toThrow();
-  });
-
   test("waitForSync should resolve when the value is uploaded", async () => {
     const { clientNode, serverNode, clientAccount } = await setupTwoNodes();