+++ examples/clerk/CHANGELOG.md
@@ -1,5 +1,31 @@
 # minimal-auth-clerk
 
+## 0.0.23
+
+### Patch Changes
+
+- jazz-react@0.8.50
+- jazz-react-auth-clerk@0.8.50
+- jazz-tools@0.8.50
+
+## 0.0.22
+
+### Patch Changes
+
+- jazz-react@0.8.49
+- jazz-react-auth-clerk@0.8.49
+- jazz-tools@0.8.49
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [635e824]
+- Updated dependencies [0a85982]
+  - jazz-tools@0.8.48
+  - jazz-react@0.8.48
+  - jazz-react-auth-clerk@0.8.48
+
 ## 0.0.20
 
 ### Patch Changes