+++ examples/password-manager/CHANGELOG.md
@@ -1,5 +1,28 @@
 # jazz-password-manager
 
+## 0.0.43
+
+### Patch Changes
+
+- jazz-react@0.8.50
+- jazz-tools@0.8.50
+
+## 0.0.42
+
+### Patch Changes
+
+- jazz-react@0.8.49
+- jazz-tools@0.8.49
+
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [635e824]
+- Updated dependencies [0a85982]
+  - jazz-tools@0.8.48
+  - jazz-react@0.8.48
+
 ## 0.0.40
 
 ### Patch Changes