+++ packages/cojson-storage-indexeddb/CHANGELOG.md
@@ -1,5 +1,29 @@
 # cojson-storage-indexeddb
 
+## 0.8.50
+
+### Patch Changes
+
+- Updated dependencies [43378ef]
+  - cojson@0.8.50
+  - cojson-storage@0.8.50
+
+## 0.8.49
+
+### Patch Changes
+
+- Updated dependencies [25dfd90]
+  - cojson@0.8.49
+  - cojson-storage@0.8.49
+
+## 0.8.48
+
+### Patch Changes
+
+- Updated dependencies [10ea733]
+  - cojson@0.8.48
+  - cojson-storage@0.8.48
+
 ## 0.8.45
 
 ### Patch Changes