+++ packages/cojson/CHANGELOG.md
@@ -1,5 +1,23 @@
 # cojson
 
+## 0.8.50
+
+### Patch Changes
+
+- 43378ef: Handle circular references in group inheritance
+
+## 0.8.49
+
+### Patch Changes
+
+- 25dfd90: Fixes the transactions collection on permissions to avoid RangeError issues
+
+## 0.8.48
+
+### Patch Changes
+
+- 10ea733: Give the ability to extend a group to accounts with reader, writer and writeOnly access level to the parent group and not only admins. The account still needs to be an admin on the child group to be able to extend it.
+
 ## 0.8.45
 
 ### Patch Changes