+++ packages/jazz-react-native-auth-clerk/CHANGELOG.md
@@ -1,5 +1,34 @@
 # jazz-react-native-auth-clerk
 
+## 0.8.50
+
+### Patch Changes
+
+- Updated dependencies [43378ef]
+  - cojson@0.8.50
+  - jazz-react-native@0.8.50
+  - jazz-tools@0.8.50
+
+## 0.8.49
+
+### Patch Changes
+
+- Updated dependencies [25dfd90]
+  - cojson@0.8.49
+  - jazz-react-native@0.8.49
+  - jazz-tools@0.8.49
+
+## 0.8.48
+
+### Patch Changes
+
+- Updated dependencies [635e824]
+- Updated dependencies [10ea733]
+- Updated dependencies [0a85982]
+  - jazz-tools@0.8.48
+  - cojson@0.8.48
+  - jazz-react-native@0.8.48
+
 ## 0.8.47
 
 ### Patch Changes